requirements.txt
python-dotenv
pandas
polars
numpy
matplotlib
seaborn
scikit-learn
xgboost
lightgbm
jupyter
shap
